- rewritten `ChipsTest` in typescript - refactored/optimized `ChipsTest.tsx`
