+++ apps/src/componentLibrary/chips/CHANGELOG.md
@@ -2,24 +2,29 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.4.0] (https://github.com/code-dot-org/code-dot-org/pull/62664)
+## [0.5.0]()
 
-* updated `Chip` and `Chips` props to include error message displayed when response required
+- rewritten `ChipsTest` in typescript
+- refactored/optimized `ChipsTest.tsx`
+
+## [0.4.0](https://github.com/code-dot-org/code-dot-org/pull/62664)
+
+- updated `Chip` and `Chips` props to include error message displayed when response required
 
 ## [0.3.1](https://github.com/code-dot-org/code-dot-org/pull/62909)
 
-* updated color variables to use `primitiveColors.css`
+- updated color variables to use `primitiveColors.css`
 
 ## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/60911)
 
-* updated `Chip` props to support native HTML Input element attributes
+- updated `Chip` props to support native HTML Input element attributes
 
 ## [0.2.0](https://github.com/code-dot-org/code-dot-org/pull/58809)
 
-* Added `gray` color
-* Added `thick` and `thin` options (chips text type)
-* Updated Chip text vertical alignment to be middle aligned
+- Added `gray` color
+- Added `thick` and `thin` options (chips text type)
+- Updated Chip text vertical alignment to be middle aligned
 
 ## [0.1.0](https://github.com/code-dot-org/code-dot-org/pull/54064)
 
-* Initial release
+- Initial release
