+++ apps/src/componentLibrary/chips/Chips.tsx
@@ -40,7 +40,7 @@ export interface ChipsProps {
  * * (✔) implementation of component approved by design team;
  * * (✔) has storybook, covered with stories and documentation;
  * * (✔) has tests: test every prop, every state and every interaction that's js related;
- * * (see apps/test/unit/componentLibrary/ChipsTest.jsx)
+ * * (see apps/test/unit/componentLibrary/ChipsTest.tsx)
  * * (?) passes accessibility checks;
  *
  * ###  Status: ```Ready for dev```
